Lightning.AI Server/requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-core-0.3.29
langchain-huggingface==0.1.2
Flask==3.1.0
transformers==4.47.1
sentence-transformers==3.3.1
accelerate==1.2.1
xformers==0.0.29
pymilvus==2.5.2
huggingface-hub==0.27.0
torch==2.5.1+cu118
Flask-Cors==5.0.0
langchain_openai-0.3.0
